+++ src/main/java/org/bytedeco/javacpp/LongPointer.java
@@ -72,8 +72,8 @@ public class LongPointer extends Pointer {
         } catch (UnsatisfiedLinkError e) {
             throw new RuntimeException("No native JavaCPP library in memory. (Has Loader.load() been called?)", e);
         } catch (OutOfMemoryError e) {
-            OutOfMemoryError e2 = new OutOfMemoryError("Cannot allocate new LongPointer(" + size + "), "
-                    + "totalBytes = " + totalBytes() + ", physicalBytes = " + physicalBytes());
+            OutOfMemoryError e2 = new OutOfMemoryError("Cannot allocate new LongPointer(" + size + "): "
+                    + "totalBytes = " + formatBytes(totalBytes()) + ", physicalBytes = " + formatBytes(physicalBytes()));
             e2.initCause(e);
             throw e2;
         }