+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 
+ * Print memory sizes in a human-readable format with `Pointer.formatBytes()`
+ * Map standard `malloc()`, `calloc()`, `realloc()`, and `free()` functions ([issue #136](https://github.com/bytedeco/javacpp/issues/136))
+
 ### November 29, 2016 version 1.2.7
  * Fix `Loader` errors that could occur due to recent changes